aws_services/un-injection/requirements.txt
aws-cdk-lib==2.94.0
constructs>=10.0.0,<11.0.0
boto3
pandas
requests
urllib3<2
openpyxl